# api-contract.yaml
openapi: 3.0.3
info:
  title: CODELESS API
  version: 0.6.0
  description: >
    API contract for the CODELESS multi-model AI chat platform.
    Features: authentication (sign-up/login/logout), user profile with
    custom avatar and preferences, multi-turn chat with AI models
    (Server-Sent Events streaming & JSON fallback), model switching
    mid-conversation, admin management of PROVIDERS (CRUD, active toggle,
    per-provider default), admin management of MODELS (CRUD, active toggle,
    platform default, provider binding), admin global token quota limit
    per user, customizable system prompt, admin user listing & management
    with chat/token stats, admin dashboard KPI metrics, forwarding chat to
    real OpenAI-compatible upstreams, and standard OpenAI-compatible endpoints
    (/v1/models, /v1/chat/completions) behind bearer auth.
    All standard JSON responses are wrapped in a unified envelope:
    { success: boolean, message: string, data: any }.
  contact:
    name: CODELESS Team
servers:
  - url: http://localhost:3000/api/v1
    description: Local development API base
  - url: https://api.codeless.example.com/api/v1
    description: Production API base (placeholder)
...
security:
  - bearerAuth: []
paths:
...
  /chat/conversations:
    get:
      tags: [Chat]
      summary: List the current user's conversations
      responses:
        '200':
          description: List of conversations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationsListApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Chat]
      summary: Start a new conversation
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConversationRequest'
      responses:
        '201':
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /chat/conversations/{conversationId}:
    parameters:
      - $ref: '#/components/parameters/ConversationId'
    delete:
      tags: [Chat]
      summary: Delete a conversation and its messages
      responses:
        '204':
          description: Conversation deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Chat]
      summary: Update conversation details (title and/or AI model)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateConversationRequest'
      responses:
        '200':
          description: Conversation updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationApiResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /chat/conversations/{conversationId}/messages:
    parameters:
      - $ref: '#/components/parameters/ConversationId'
    get:
      tags: [Chat]
      summary: Get the message history of a conversation
      responses:
        '200':
          description: Ordered list of messages (oldest first)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesListApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Chat]
      summary: Send a message and receive the model's reply
      description: >
        Streamed as `text/event-stream`: one `event: token` per generated token,
        followed by a final `event: done` carrying the saved message id.
        Clients that send `Accept: application/json` instead get the complete
        answer in a single non-streamed response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '200':
          description: Model response (streamed or complete, depending on Accept header)
          content:
            text/event-stream:
              schema:
                type: string
                format: binary
              example: |
                event: token
                data: {"content":"Hi"}

                event: token
                data: {"content":"!"}

                event: done
                data: {"messageId":"5b1e2b2a-df3a-4e9a-9a11-1a2b3c4d5e6f"}
            application/json:
              schema:
                $ref: '#/components/schemas/MessageApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ConversationId:
      name: conversationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    CreateConversationRequest:
      type: object
      properties:
        modelId:
          type: string
          format: uuid
          description: Optional — falls back to the platform default model if omitted.
        title:
          type: string
    UpdateConversationRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          description: Updated title for the conversation
          example: "Updated Conversation Name"
        modelId:
          type: string
          format: uuid
          description: >
            Switch the model that answers this conversation from now on.
            Must be an ACTIVE model whose provider is also ACTIVE.
            At least one of title/modelId is required.
    Conversation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        modelId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SendMessageRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        conversationId:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
...
    Error:
      type: object
      required: [success, message, data]
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Invalid email or password"
        data:
          nullable: true
          example: null
        statusCode:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
    ApiResponse:
      type: object
      required: [success, message, data]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation successful"
        data:
          nullable: true
          description: Response payload
...
    ConversationApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Conversation'
    ConversationsListApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Conversation'
    MessageApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Message'
    MessagesListApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Message'
...